requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cmdstanpy==1.2.5
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
dotenv==0.9.9
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fonttools==4.58.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.16.0
holidays==0.74
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lml==0.2.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.41.0
numpy==2.2.6
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
prophet==1.1.7
protobuf==6.31.0
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==20.0.0
pydeck==0.9.1
pyexcel-io==0.6.7
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
stanio==0.5.1
streamlit==1.45.1
tenacity==9.1.2
texttable==1.7.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
xgboost==3.0.2
xlwt==1.3.0
